+++ stroom-app/src/test/java/stroom/data/store/impl/fs/TestOrphanFileFinder.java
@@ -91,7 +91,11 @@ class TestOrphanFileFinder extends AbstractCoreIntegrationTest {
 
         assertThat(Files.exists(test1)).isTrue();
 
-        final List<String> fileList = scan();
+        final FsOrphanFileFinderSummary summary = new FsOrphanFileFinderSummary();
+        final List<String> fileList = scan(summary);
+        assertThat(summary.toString()).isEqualTo("Summary:\n" +
+                "RAW_EVENTS\n" +
+                " - 2021/10/18 - 1\n");
         final List<FsVolume> volumeList = volumeService.find(FindFsVolumeCriteria.matchAll()).getValues();
         assertThat(volumeList.size()).isEqualTo(1);
         assertThat(fileList).contains(FileUtil.getCanonicalPath(test1));
@@ -99,7 +103,9 @@ class TestOrphanFileFinder extends AbstractCoreIntegrationTest {
 
     @Test
     void testScan() throws IOException {
-        scan();
+        final FsOrphanFileFinderSummary summary = new FsOrphanFileFinderSummary();
+        scan(summary);
+        assertThat(summary.toString()).isEqualTo("");
 
         final ZonedDateTime oldDate = ZonedDateTime.now(ZoneOffset.UTC).plusDays(NEG_SIXTY);
 
@@ -165,7 +171,11 @@ class TestOrphanFileFinder extends AbstractCoreIntegrationTest {
                     ZonedDateTime.now(ZoneOffset.UTC).plusDays(NEG_FOUR).toInstant().toEpochMilli());
 
             // Run the clean
-            final List<String> fileList = scan();
+            final FsOrphanFileFinderSummary summary2 = new FsOrphanFileFinderSummary();
+            final List<String> fileList = scan(summary2);
+            assertThat(summary2.toString()).isEqualTo("Summary:\n" +
+                    "RAW_EVENTS\n" +
+                    " - 2021/10/18 - 3\n");
 
             final List<FsVolume> volumeList = volumeService.find(FindFsVolumeCriteria.matchAll()).getValues();
             assertThat(volumeList.size()).isEqualTo(1);
@@ -205,7 +215,9 @@ class TestOrphanFileFinder extends AbstractCoreIntegrationTest {
                 .as("Must be saved to at least one volume")
                 .isTrue();
 
-        scan();
+        final FsOrphanFileFinderSummary summary = new FsOrphanFileFinderSummary();
+        scan(summary);
+        assertThat(summary.toString()).isEqualTo("");
 
         files = fileFinder.findAllStreamFile(meta);
 
@@ -217,12 +229,16 @@ class TestOrphanFileFinder extends AbstractCoreIntegrationTest {
                 .as("Volumes should still exist as they are new")
                 .isTrue();
 
-        scan();
+        final FsOrphanFileFinderSummary summary2 = new FsOrphanFileFinderSummary();
+        scan(summary2);
+        assertThat(summary2.toString()).isEqualTo("");
     }
 
     @Test
     void testScanLotsOfFiles() throws IOException {
-        scan();
+        final FsOrphanFileFinderSummary summary = new FsOrphanFileFinderSummary();
+        scan(summary);
+        assertThat(summary.toString()).isEqualTo("");
 
         final String feedName = FileSystemTestUtil.getUniqueTestString();
         final long endTime = System.currentTimeMillis();
@@ -240,13 +256,16 @@ class TestOrphanFileFinder extends AbstractCoreIntegrationTest {
             }
         }
 
-        scan();
+        final FsOrphanFileFinderSummary summary2 = new FsOrphanFileFinderSummary();
+        scan(summary2);
+        assertThat(summary2.toString()).isEqualTo("");
     }
 
-    private List<String> scan() {
+    private List<String> scan(final FsOrphanFileFinderSummary summary) {
         final List<String> fileList = new ArrayList<>();
         final Consumer<Path> orphanConsumer = path -> {
             fileList.add(FileUtil.getCanonicalPath(path));
+            summary.addPath(path);
         };
         fsOrphanFileFinderExecutor.scan(orphanConsumer, new SimpleTaskContext());
         return fileList;